Move persistent state associated with ShouldInjectFieldsMatcher to its own class
